Continued camera servo turret

+++ include/diffdrive_arduino/arduino_comms.h
@@ -22,7 +22,7 @@ public:
   void readEncoderValues(int &val_1, int &val_2);
   void setMotorValues(int val_1, int val_2);
   void setPidValues(float k_p, float k_d, float k_i, float k_o);
-  void setServoValues(int pos_array[]);
+  void setServoValues(double pos_array[]);
   // void readIMUValues(int &val_1, int &val_2);
   // void readServoValues(int &val_1, int &val_2);
